backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pdfplumber>=0.10.0
openai>=1.3.0
loguru>=0.7.0
python-dotenv>=1.0.0
supabase>=2.0.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
gunicorn>=21.0.0
